bookmark api 피드백 반영
