+++ article/views/__init__.py
@@ -2,6 +2,6 @@ from .article_list_create_api_view import ArticleListCreateAPIView
 from .article_retrieve_update_destroy_api_view import (
     ArticleRetrieveUpdateDestroyAPIView,
 )
-from .bookmark_list_create_api_view import BookmarkListCreateAPIView
+from .bookmark_create_api_view import BookmarkCreateAPIView
 from .bookmark_destroy_api_view import BookmarkDestroyAPIView
